refactor: user resolver pascal casing
