+++ packages/hash/api/src/graphql/resolvers/index.ts
@@ -29,7 +29,7 @@ import {
 } from "./entity/impliedHistory";
 
 import { me } from "./knowledge/user/me";
-import { isShortnameTaken } from "./knowledge/user/isShortnameTaken";
+import { isShortnameTaken } from "./knowledge/user/is-shortname-taken";
 import { deprecatedCreateEntityType } from "./entityType/createEntityType";
 import { SYSTEM_TYPES, SystemType } from "../../types/entityTypes";
 import { entityTypeTypeFields } from "./entityType/entityTypeTypeFields";
@@ -115,7 +115,7 @@ import { deletePersistedComment } from "./knowledge/comment/delete";
 import { updatePersistedCommentText } from "./knowledge/comment/update-text";
 import { blockChildEntity } from "./knowledge/block/data-entity";
 import { loggedInAndSignedUpHashInstanceAdmin } from "./middlewares/loggedInAndSignedUpHashInstanceAdmin";
-import { createUser } from "./knowledge/user/createUser";
+import { createUser } from "./knowledge/user/create-user";
 
 /**
  * @todo: derive these from the statically declared system type names
