+++ packages/hash/api/src/graphql/resolvers/knowledge/user/is-shortname-taken.ts
